SCA feedback pass,
Fixed warnings and some SCA feedback issues. Tested good on map with
large laggy ship, not tested with cargo ships which was causing error
mentioned in previous commit.

+++ Sources/Sandbox.Game/Game/Entities/Cube/MyTerminalBlock.cs
@@ -302,7 +302,7 @@ namespace Sandbox.Game.Entities.Cube
                 OffsetText = true,
                 TargetMode = GetPlayerRelationToOwner(),
                 Entity = this,
-                BlinkingTime = allowBlink && IsBeingHacked ? MyGridConstants.HACKING_INDICATION_TIME_MS / 1000 : 0
+                BlinkingTime = allowBlink && IsBeingHacked ? (float)MyGridConstants.HACKING_INDICATION_TIME_MS / 1000 : 0
             });
 
             return m_hudParams;
